(improvement)(semantic) support add priorExts in llmParser
